Designing a decent data model.
Since QCP copies the data it uses to make plots, we need a good strategy
to connect the datasets to these copies. A possibility would be to use
signals and slots to reload the data associated to a curve whenever the
associated dataset changes. To this aim, I will use a centralised
DataManager that stores all the datasets and connect/disconnect them to
their associated curves. DataManager will be also responsible for the
Dataset creation, making DataFactory unnecessary.