Removed modal dialogs in favour of non-modal windows (as in xmgrace)
1) A non-modal dialog is a dialog that leave the main window usable. In
this way more than one dialog can be open at the same time (cfr.
xmgrace). Nearly all disgrace dialogs will be non-modal
2) Added an "apply" button to the dialogs. In this way it will be
possible to apply changes without having to close the dialog.
3) apply/ok clicked() signals will be connected to MainWindow slots. The
idea is that all the "business" logic that deals with the plot
appearance and datasets should be clustered together and not scattered
all around different dialogs. In this way it will be easier to make
disgrace scriptable. I don't know if MainWindow is the right object to
contain all this business logic, but it won't be hard to move it
somewhere else in the futuer.